Solve tab view measuring issue. Closes #673.
