+++ library/res/layout/abs__action_bar_tab_bar_view.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 
-<com.actionbarsherlock.internal.widget.IcsLinearLayout
+<com.actionbarsherlock.internal.widget.TabsLinearLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
     style="?attr/actionBarTabBarStyle"
 />
