+++ library/src/com/actionbarsherlock/internal/widget/ScrollingTabContainerView.java
@@ -186,7 +186,7 @@ public class ScrollingTabContainerView extends NineHorizontalScrollView
     }
 
     private IcsLinearLayout createTabLayout() {
-        final IcsLinearLayout tabLayout = (IcsLinearLayout) LayoutInflater.from(getContext())
+        final TabsLinearLayout tabLayout = (TabsLinearLayout) LayoutInflater.from(getContext())
                 .inflate(R.layout.abs__action_bar_tab_bar_view, null);
         tabLayout.setMeasureWithLargestChildEnabled(true);
         tabLayout.setLayoutParams(new LinearLayout.LayoutParams(
